+++ downstream-test/src/test/java/react4j/downstream/BuildStatsTest.java
@@ -71,6 +71,13 @@ public class BuildStatsTest
     compareSizesForBranch( "webspeechdemo", false );
   }
 
+  @Test
+  public void heartRateMonitor()
+    throws Exception
+  {
+    compareSizesForBranch( "heart-rate-monitor", false );
+  }
+
   private void compareSizesForBranch( @Nonnull final String branch, final boolean includeBranchInFixtureKey )
     throws Exception
   {