+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Add the `react4j-heart-rate-monitor` to the variants tracked as part of the release process.
+
 ### [v0.172](https://github.com/react4j/react4j/tree/v0.172) (2020-07-23) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.171...v0.172) · [API Differences](https://react4j.github.io/api-diff?key=react4j&old=0.171&new=0.172)
 
 The release includes 1 non breaking API change and 1 breaking API change.